- 修复前端写blog的bug - 增加后端分类和标签功能(待测试)

+++ backend/src/main/java/com/lilac/domain/dto/article/ArticleAddRequest.java
@@ -48,5 +48,5 @@ public class ArticleAddRequest {
     /**
      * 标签列表
      */
-    private List<Integer> tagIds;
+    private List<Long> tagIds;
 }
